add hiccup_utilities.py
